requirements.txt
fastapi
uvicorn
openai
faiss-cpu
numpy
python-docx
pytesseract
Pillow
PyMuPDF
jinja2
python-multipart
python-dotenv
